ai_models/hazard_traversability/requirements.txt
numpy
matplotlib
pyyaml
datasets
torch
torchvision
pillow
